requirement.txt
streamlit==1.34.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
numpy==1.26.4
scikit-learn==1.4.2
